requirements.txt
chamferdist==1.0.0
imageio==2.19.3
kornia==0.6.4
matplotlib==3.5.1
numpy==1.21.5
oflibpytorch==2.1.0
opencv_python==4.5.5.64
Pillow==9.4.0
pypng==0.20220715.0
pytz==2022.1
scipy==1.7.3
setuptools==61.2.0
torchmetrics==0.11.0
torchsummary==1.5.1
tqdm==4.64.0
